django/requirements.txt
absl-py==1.1.0
asgiref==3.5.2
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Django==4.0.4
django-environ==0.8.1
django-storages==1.12.3
flatbuffers==1.12
gast==0.4.0
google-api-core==2.8.1
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.0
google-cloud-secret-manager==2.11.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpc-google-iam-v1==0.12.4
grpcio==1.46.3
grpcio-status==1.46.3
h11==0.13.0
h5py==3.7.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
nltk==3.7
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
proto-plus==1.20.5
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF2==2.0.0
PySocks==1.7.1
python-dotenv==0.20.0
PyVirtualDisplay==3.0
regex==2022.6.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
webdriver-manager==3.7.0
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
zipp==3.8.0
